The Algorithm_Factory has this logic on looking for an object:
 - Check the cache; if found, return value
 - Populate cache, if the value is already there, delete the old
   object and save the new one.
 - Recheck the cache value

Raja <kalx91@gmail.com> pointed out on the list that this could race
if multiple threads called a lookup function in close succession while
the cache was cold. All of them would fail the lookup, then each of
them would add it, but the values returned would be deleted by other
threads.

Instead, declare that first write wins. Then, the cache stays
consistent even if there is a race, the only issue is an extra search
and delete.

Modify GOST and Skein, as their name() function did not roundtrip
properly which caused failures otherwise.