Full support for renegotiation including RFC 5746 extensions for client and server. Server side can handle SCSV values as well, client always sends the extension instead.
Handle an empty SNI extension coming back from the server - this is
used to indicate that it understood the name. Also add better checking
for extensions by passing in what the supposed size of the extension
is.

Only send the secure negotiation extension in the server hello if the
client indicated support for it.