Deleting the return of private_key_for in the TLS server forces the credentials server to return a new copy each time which is slow and mostly pointless. Instead, specify that the key remains owned by the credentials manager.
This is theoretically an issue if you have thousands of keys to
manage; the credentials server doesn't actually know when they have
gone out of scope until its destructor runs. So it could be forced to
use a lot of memory in the meantime. I'm not sure that this is a case
worth optimizing for, at least until someone comes along who actually
has this as a problem.