Add anonymous DH/ECDH ciphersuites to the cipher list. Interop checked against OpenSSL.
One big issue that needs to be resolved is that with these
ciphersuites available to be negotiated, we want to make sure they
only are used when the application/user expects them to.  Problem is
that PSK and SRP are "anonymous" but authenticated via the shared
secret. We need to be able to distinguish these on a policy
level. Otherwise a MITM could simply offer anon DH, which would be
somewhat unfortunate. A client could detect this in the handshake
callback, but might not.

In the short term to ensure this doesn't occur, disable both anon DH
and PSK/SRP in the default policy.