Expose Handshake_Message as an API type.
Use override as appropriate in tls_messages.h

Add a callback to Handshake_State which, if set, is called for each
Handshake_Message& as it is received or sent. Not yet exposed at the
application level.

Use default values in client and server hellos to minimize the need
for explicit settings in constructors.