Use a.data() instead of &a[0] for containers
When running a program which was compiled by MSVC in debug mode,
vectors and strings complain with an assertion if elements which are out
of bounds are accessed. In the Botan sources, the idiom &a[0] is
frequently used to get the address of the underlying raw data of a
container. This clashes with the above behavior in the case that the
container is empty. The alternative syntax a.data() works just fine.

For this patch, I first ran the following command:
find . -type f -print0 \
| xargs -0 sed -ri 's/&([a-zA-Z_]+)\[0\]/\1.data()/g'

Afterwards, I manually reviewed all changes, reverting false positives
and other inappropriate changes.