Force semicolons at the end of BOTAN_REGISTER_* macro invocations
All BOTAN_REGISTER_* macros are defined as

    namespace { some_command(); }

So, if such a macro is used with a semicolon at the end, we have
`namespace { ... };` which is unnecessary and makes gcc complain when
run with with -Wpedantic.

However, for consistency, it is great to end those macro invocations
with a semicolon. This commit forces semicolons by appending a dummy
definition with the necessary semicolon missing.