In RSA, check that the input is less than the modulus n before blinding rather than after.  After blinding the value is always reduced mod n so the condition is never met.
This may be the cause of RSA test failures described in GH #174

The scenario was that during randomized corruption tests we
occasionally provide an input which was greater than the modulus.
When that happened the value was effectively reduced mod n, so the
self-check would later fail, because the decrypted result (reduced mod n)
would be compared with the original (larger than n) input.