Remove RNG::gen_mask, which had undefined behavior when bits >= 32
Bug found by Daniel Neus

The function wasn't being used anywhere in the library (and was only
added in 1.11.20) so it seems easier to remove than fix. And removing
it serves to put any user on notice that something bad happened;
Daniel tested this as returning just 0 when bits >= 32 with his
system's compiler.