If EPERM when writing to the random device, return silently.
That failure seems like an explicit system policy decision, so accept
it as such. Versus say EBADF or EFAULT which would instead suggest a
library bug.

This return is seen on OS X Travis CI, unclear if this is a problem
with OS X or with how Travis does OS X containers.