Fix heap overflow in ECC point multiplication
If affine coordinates larger than the prime modulus were given,
a later multiplication could overflow the size of an allocated
output buffer, which was sized based on the size of the prime.

This will cause an overflow into either the system heap or if the
mlock/mmap pool allocator is in use, then into the adjacent key
material stored in the pool.

Reported by Alex Gaynor who found it with AFL

Also fix a one word overwrite in P-521 reduction. Found with AFL