Fix two bugs in 1.11.27 which caused test failures.
The check on each individual size in curve_mul is too strict since
we rely on redc(x*1) during the on the curve computation.

Fix an off by one in ressol which caused it to occasionally reject
valid values.

Updating version 1.11.28 since existing 1.11.27 tag already pushed :(

Fix an off-by-one in ressol which would cause it to occasionly
give up too early.