Add KDF1 as explicit dependency in MCEIES
Previously MCEIES used KEM with a raw SHA-512 hash. When the KDF
oriented KEM interface was added in 72f0f0ad2a it switched to using
KDF1(SHA-512) since for a 64-byte output and no salt, the output
matches the original hash. This avoids breaking the format.

Prevents runtime failures when MCEIES is enabled in the build but KDF1
is not. GH #369