Remove configure extra information from build.h
Removes user and hostname: as mentioned by a downstream maintainer, the
information should not be included in packages. In our projects, the
information is contained in the amalgamation check-in commit message.

Removes timestamp: this ensures that the header content is equal in two
equal configure runs.

Both changes together make it possible to effectively use compiler
caches for Botan (ccache on Unix, in direct mode; clcache on Windows).