RNG cleanups
Removes implicit references to the entropy sources.

Chagnes Stateful_RNG to have pointers to an optional RNG and an optional
Entropy_Sources which it can pull from. If a fork is detected and there
is no seed source set an exception is thrown.

Officially deprecates both X9.31 and HMAC_RNG. It would be better to
consolidate on one userspace RNG design.

Test uses a short string for repeatable RNG outputs. It now hashes
the input with SHA-512 to produce the DRBG seed.